requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyascii==0.3.2
asarPy==1.0.1
async-timeout==5.0.1
attrs==25.1.0
audioread==3.0.1
babel==2.16.0
bangla==0.0.2
bcc==0.18.0
blinker==1.9.0
blis==1.2.0
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
Brlapi==0.8.3
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.1
chrome-gnome-shell==0.0.0
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.4
command-not-found==0.3
confection==0.1.5
contourpy==1.0.7
coqpit==0.0.17
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
cymem==2.0.11
Cython==3.0.11
dateparser==1.1.8
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
distlib==0.3.9
distro==1.7.0
docopt==0.6.2
einops==0.8.0
encodec==0.1.1
filelock==3.17.0
Flask==3.1.0
fonttools==4.38.0
frozenlist==1.5.0
fsspec==2024.12.0
g2pkk==0.1.2
grpcio==1.70.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
hangul-romanize==0.1.0
hidpidaemon==18.4.6
httplib2==0.20.2
huggingface-hub==0.28.0
idna==3.10
importlib-metadata==4.6.4
inflect==7.5.0
itsdangerous==2.2.0
jamo==0.4.1
jeepney==0.7.1
jieba==0.42.1
Jinja2==3.1.5
joblib==1.4.2
jsonlines==1.2.0
kernelstub==3.1.4
keyring==23.5.0
kiwisolver==1.4.4
langcodes==3.5.0
language-selector==0.1
language_data==1.3.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
librosa==0.10.0
llvmlite==0.43.0
louis==3.20.0
macaroonbakery==1.3.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.7.0
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.12
mutagen==1.47.0
netaddr==0.8.0
netifaces==0.11.0
networkx==2.8.8
nltk==3.9.1
num2words==0.5.14
numba==0.60.0
numpy==1.22.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
openai-whisper==20240930
packaging==24.2
pandas==1.5.3
Pillow==9.4.0
platformdirs==4.3.6
pooch==1.8.2
preshed==3.0.9
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pydbus==0.6.0
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynndescent==0.5.13
pyparsing==2.4.7
pypinyin==0.53.0
pyRFC3339==1.1
pysbd==0.3.4
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
python-apt==2.4.0+ubuntu4
python-crfsuite==0.9.11
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-gnupg==0.4.8
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.2
regex==2024.11.6
repolib==2.2.1
repoman==1.4.0
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.11.4
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
semantic-version==2.10.0
sessioninstaller==0.0.0
setuptools-rust==1.10.2
shellingham==1.5.4
shiboken6==6.7.2
six==1.16.0
smart-open==7.1.0
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
SudachiDict-core==20241021
SudachiPy==0.6.10
sympy==1.13.1
systemd-python==234
tensorboard==2.18.0
tensorboard-data-server==0.7.2
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1
torchaudio==2.5.1
tqdm==4.67.1
trainer==0.0.36
transformers==4.48.1
triton==3.1.0
TTS @ git+https://github.com/coqui-ai/TTS.git@dbf1a08a0d4e47fdad6172e433eeb34bc6b13b4e
typeguard==4.4.1
typer==0.15.1
typing_extensions==4.12.2
tzlocal==5.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
umap-learn==0.5.7
Unidecode==1.3.8
urllib3==2.3.0
vboxapi==1.0
virtualenv==20.29.1
wadllib==1.3.6
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
xdg==5
xkit==0.0.0
yarl==1.18.3
zipp==1.0.0
